requirements.txt
absl-py==0.11.0
click~=7.1.2
pandas==1.1.1
scikit-learn==0.24.0
tensorboard-plugin-profile==2.4.0
tensorflow==2.4.0
tensorflow-text==2.4.3
tf-models-official==2.4.0
transformers==4.1.1
torch===1.7.1
numpy~=1.19.5